view-report - Move ReportView into app/_components
